Expand tests for inlining to a breaking point.
Removing access widening from SuperAccessors has broken
this test. While Inliners assumes that `B#xvar` is public,
it generates a call to the original name, rather than to
the expanded name.
