+++ test/files/run/inline-access-levels/A_1.scala
@@ -1,0 +1,13 @@
+package test
+
+object A {
+  private[test] var packagePrivate: Int = -1
+}
+
+trait A1 {
+  private[test] var packagePrivateClass: Int = -1
+}
+
+class A extends A1 {
+//  private[test] var packagePrivateClass: Int = -1
+}
